integration bdd

+++ src/main/java/fr/eni/filmo/bll/AvisService.java
@@ -8,5 +8,5 @@ import fr.eni.filmo.bo.Avis;
 public interface AvisService {
 	public void insert(Avis a);
 	public List<Avis> selectAll();
-	public Avis select(int id);
+	public Avis select(Long id);
 }
