+++ src/main/java/fr/eni/filmo/ihm/StringToPersonneConverter.java
@@ -19,7 +19,7 @@ public class StringToPersonneConverter implements Converter<String, Personne>{
 	}
 	@Override
 	public Personne convert(String source) {
-		Integer theid = Integer.parseInt(source);
+		Long theid = Long.parseLong(source);
 		try {
 			return service.select(theid);
 		} catch (PersonneNotFound e) {
